+++ types/disruptiveTechnologies/humidity/uplink.spec.js
@@ -1,0 +1,58 @@
+const chai = require("chai");
+const { validate } = require("jsonschema");
+const rewire = require("rewire");
+const utils = require("test-utils");
+
+const { assert } = chai;
+
+describe("Digital Technologies Humidity Sensor Uplink", () => {
+  let humiditySchema = null;
+  let consume = null;
+  before((done) => {
+    const script = rewire("./uplink.js");
+    consume = utils.init(script);
+    utils
+      .loadSchema(`${__dirname}/humidity.schema.json`)
+      .then((parsedSchema) => {
+        humiditySchema = parsedSchema;
+        done();
+      });
+  });
+
+  describe("consume()", () => {
+    it("should decode the Digital Technologies Humidity Sensor payload", () => {
+      const data = {
+        data: {
+          event: {
+            eventId: "c505kmuj0aoraraqu5g0",
+            targetName:
+              "projects/c3t7p26j4a2g00de1sng/devices/emuc4ah9r13um94o4pp3hdg",
+            eventType: "humidity",
+            data: {
+              humidity: {
+                temperature: 22.45,
+                relativeHumidity: 17,
+                updateTime: "2021-05-16T06:13:46.369000Z",
+              },
+            },
+            timestamp: "2021-09-14T08:16:27.517331Z",
+          },
+          labels: { name: "Temperature Simulator" },
+        },
+      };
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.topic, "humidity");
+        assert.equal(value.data.temperature, 22.45);
+        assert.equal(value.data.humidity, 17);
+
+        validate(value.data, humiditySchema, { throwError: true });
+      });
+
+      consume(data);
+    });
+  });
+});